% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{createData}
\alias{createData}
\title{Extract the info necessary for the gibbs sampler.}
\usage{
createData(
  A,
  B,
  name.y,
  PIVs,
  regression.formula,
  pmatch.formula,
  family.Y1,
  family.Y0
)
}
\arguments{
\item{A}{x}

\item{B}{x}

\item{name.y}{x}

\item{PIVs}{x}

\item{regression.formula}{x}

\item{pmatch.formula}{x}

\item{family.Y1}{x}

\item{family.Y0}{x}

\item{data}{The observed data consisting of. Must be created with the ... function.}
}
\value{
List of prior distributions.
}
\description{
Extract the info necessary for the gibbs sampler.
}
\author{
Michel H Hof, \email{m.h.hof@amsterdamumc.nl}
}
