% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{createPriors}
\alias{createPriors}
\title{Set the priors for the Gibbs sampler. Possible to modify afterwards if necessary (see example)}
\usage{
createPriors(data)
}
\arguments{
\item{data}{The observed data consisting of. Must be created with the ... function.}
}
\value{
List of prior distributions.
}
\description{
Set the priors for the Gibbs sampler. Possible to modify afterwards if necessary (see example)
}
\author{
Michel H Hof, \email{m.h.hof@amsterdamumc.nl}
}
