% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{createStart}
\alias{createStart}
\title{Create start values for the Gibbs sampler}
\usage{
createStart(data, chain = NULL)
}
\arguments{
\item{data}{The observed data consisting of. Must be created with the ... function.}

\item{chain}{Previous run of the Gibbs sampler.}
}
\value{
List of starting values.
}
\description{
Can be modified afterwards if necessary. RANDOMNESS!
}
\author{
Michel H Hof, \email{m.h.hof@amsterdamumc.nl}
}
