% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\name{runGibbs}
\alias{runGibbs}
\title{Main function of the StEM estimator.}
\usage{
runGibbs(data, priors, start, n.iter, w)
}
\description{
The data mut be prepared with the function \link{createData}. Two types of regression models can be fitted to the data
linear regression and logistic regression models. Note that this version of the StEM algorithm is very limited and can only fit the model from the simulation study;
i.e. the PIVs are assumed to be ACAR/DCAR and that the outcome variable is assumed to be normally distributed in observations from
file B without a match in file A.
}
\author{
Michel H Hof, \email{m.h.hof@amsterdamumc.nl}
}
